+++ changelog.md
@@ -2,6 +2,15 @@
 
 ## Version 7
 
+### 7.0.9
+- Adding Route `getPosRecurrences`
+    - Available since HDP 4.2.10
+    - POST `/nitro/projects/{project_ID}/datasets/{dataset_ID}/pos/recurrences`
+
+- Adding Route `updatePosRecurrence`
+    - Available since HDP 4.2.10
+    - POST `/nitro/projects/{project_ID}/datasets/{dataset_ID}/pos/{pos_ID}/recurrences`
+
 ### 7.0.8
 - Using string delimiters when creating datasets in mongo